Decide on name and update scenes
I will go with a Tower Defence game this year.
You can choose between playing the Life or the Death Side.
Basically Angels vs. Demons.

The game is going to be called Life Death Tower Defence.
I want to have the Tower on one line, defence below it.
Vertically, the Life will overlap with Tower on e and Death on t.
All in all caps.

To only see one scene at a time I've updated the scene components and
introduced a new initialState property alongside an action creator and
reducer.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>